+++ src/mosaic/__init__.py
@@ -2,6 +2,10 @@
 """.. include:: ../../README.md"""
 __all__ = ['Clustering', 'GridSearchCV', 'Similarity', 'UMAPSimilarity']
 
+METRICS = {'correlation', 'NMI', 'JSD', 'GY'}
+MODES = {'CPM', 'modularity', 'linkage', 'kmedoids'}
+NORMS = {'joint', 'geometric', 'arithmetic', 'min', 'max'}
+
 import mosaic.utils  # noqa: F401
 from .clustering import Clustering
 from .gridsearch import GridSearchCV