+++ src/mosaic/__init__.py
@@ -13,4 +13,4 @@ from .similarity import Similarity
 from .umap_similarity import UMAPSimilarity
 
 
-__version__ = '0.3.0'
+__version__ = '0.3.1'