+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and [Element](https://github.com/vector-im/element-android)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.3.1] - 2022-08-04
 ### Bugfix 🐛:
 - Fixing floating-point error for `np.float16, np.float32` in `Similarity().fit()`, see#9
 
@@ -124,7 +127,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release
 
 
-[Unreleased]: https://github.com/moldyn/MoSAIC/compare/v0.3.0...main
+[Unreleased]: https://github.com/moldyn/MoSAIC/compare/v0.3.1...main
+[0.3.1]: https://github.com/moldyn/MoSAIC/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/moldyn/MoSAIC/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/moldyn/MoSAIC/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/moldyn/MoSAIC/compare/v0.2.0...v0.2.1